Feature/repos (#4)
* feat: introduce API and database migrator

This commit sets up a new API and a database migrator workspace. Dependencies for both projects are specified in their respective Cargo.toml files, and a SQL migration script is added for creating the `repos` table.

* chore: update dependencies in Cargo.lock

Upgraded versions for flate2, hyper-util, libc, and redox_syscall. Removed unused dependencies including pin-project, its internal components, and tower-related packages. This streamlines our dependencies and resolves potential security or stability issues.

* feat: update Dockerfile mount sources

Change bind mount sources from 'src' to 'api' and add 'migrator'. This allows for more modular and organized source directories within the container build process.

* feat: add support for config files in database

* feat: add support for creating multiple cargo repositories

* feat: add migrator binary to Docker image

Include the migrator binary in the Docker image by copying it during the build process. Adjust the Dockerfile to ensure the migrator binary is available in the final runtime stage.

* feat: add migration directory support in Docker setup

Added a step in Dockerfile to copy migration files and set an environment variable for the migrations directory. Modified the migrator to use the environment variable for the migrations path.

* feat: add support for creating, listing and deleting repos