Open source improvements (#15)
* fix: remove catalog-info and update GitHub workflow branches

Deleted the unused `catalog-info.yaml` file as it is no longer needed. Updated the GitHub workflow to trigger on pushes to the "main" and "develop" branches, ensuring builds are consistent across primary branches.

* Create LICENSE

* feat: add Contributor Covenant Code of Conduct

This new document outlines our pledge for a harassment-free experience and defines behavioral standards. It details enforcement responsibilities and consequences for inappropriate behavior to maintain a healthy community.

* docs: add setup instructions and update authority URL

Added detailed setup instructions in the README.md, including steps for configuring the database, running migrations, and starting both the API and UI. Also, updated the authority URL in config.toml to use a local server for authentication.